Only give access to draft if user has edit permissions on dataset
