+++ packages/openneuro-app/src/scripts/authentication/profile.js
@@ -14,3 +14,16 @@ export const getProfile = () => {
   const accessToken = cookies.get('accessToken')
   return accessToken ? parseJwt(accessToken) : null
 }
+
+// Return true if the active user has write permission
+export const hasEditPermissions = (permissions, userId) => {
+  if (userId) {
+    const permission = permissions.find(perm => perm.user.id === userId)
+    return (
+      (permission &&
+        (permission.level === 'admin' || permission.level === 'rw')) ||
+      false
+    )
+  }
+  return false
+}
